revert major refactor, just extend existing class with solr methods
